+++ src/komm/_lossless_coding/ShannonCode.py
@@ -36,7 +36,7 @@ class ShannonCode(FixedToVariableCode):
          (1,): (1, 0, 0, 0),
          (2,): (1, 0, 0, 1)}
         >>> code.rate(pmf)  # doctest: +FLOAT_CMP
-        np.float64(1.6)
+        1.6
 
         >>> code = komm.ShannonCode(pmf, 2)
         >>> code.enc_mapping
@@ -50,7 +50,7 @@ class ShannonCode(FixedToVariableCode):
          (2, 1): (1, 1, 0, 0, 0, 1, 0),
          (2, 2): (1, 1, 0, 0, 0, 1, 1)}
         >>> code.rate(pmf)  # doctest: +FLOAT_CMP
-        np.float64(1.1)
+        1.1
     """
 
     def __init__(self, pmf: npt.ArrayLike, source_block_size: int = 1):
